Added CRUD functionality to the DepartmentController

+++ src/main/java/com/leverx/employeestat/rest/service/impl/DepartmentServiceImpl.java
@@ -1,16 +1,15 @@
 package com.leverx.employeestat.rest.service.impl;
 
 import com.leverx.employeestat.rest.entity.Department;
+import com.leverx.employeestat.rest.exception.DuplicateDepartmentException;
 import com.leverx.employeestat.rest.exception.RecordNotFoundException;
 import com.leverx.employeestat.rest.repository.DepartmentRepository;
 import com.leverx.employeestat.rest.service.DepartmentService;
-import java.util.List;
-import java.util.UUID;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import javax.transaction.Transactional;
+import java.util.List;
+import java.util.UUID;
 
 @Service
 public class DepartmentServiceImpl implements DepartmentService {
@@ -25,19 +24,31 @@ public class DepartmentServiceImpl implements DepartmentService {
     @Override
     @Transactional
     public Department save(Department department) {
+        if (departmentRepository.existsByName(department.getName())) {
+            throw new DuplicateDepartmentException("Department with name=" + department.getName() + " already exists");
+        }
         return departmentRepository.save(department);
     }
 
     @Override
     @Transactional
     public void deleteById(UUID id) {
+        if (!departmentRepository.existsById(id)) {
+            throw new RecordNotFoundException("Department with id=" + id + " not found for deleting");
+        }
         departmentRepository.deleteById(id);
     }
 
     @Override
     @Transactional
     public Department update(Department department) {
-        return departmentRepository.save(department);
+        if (departmentRepository.existsById(department.getId())) {
+            return departmentRepository.save(department);
+        } else if (!departmentRepository.existsByName(department.getName())) {
+            return departmentRepository.save(department);
+        } else {
+            throw new DuplicateDepartmentException("Department with name=" + department.getName() + " already exists");
+        }
     }
 
     @Override
@@ -46,12 +57,13 @@ public class DepartmentServiceImpl implements DepartmentService {
         return departmentRepository.findAll();
     }
 
-
     @Override
     @Transactional
     public Department getById(UUID id) {
         return departmentRepository.findById(id)
-                .orElseThrow(() -> {throw new RecordNotFoundException("Department with id=" + id + " not found");});
+                .orElseThrow(() -> {
+                    throw new RecordNotFoundException("Department with id=" + id + " not found");
+                });
     }
 
     @Override
